Use correct variable types.

+++ hd44780.h
@@ -1,8 +1,10 @@
 #ifndef LCD_H
 #define LCD_H
 
+#include <stdint.h>  // For uint8_t.
+
 void lcd_print(const char*);
-void lcd_print_position(int, int, const char*);
+void lcd_print_position(uint8_t, uint8_t, const char*);
 void hd44780_setup();
 
 #endif
